+++ src/ts/binaryen_wasm_ts.js
@@ -1,0 +1,4 @@
+/* the purpose of this file is simply to enable compilation of the binaryen-wrapper.ts file.
+Compilation of the binaryen_ts target creates the real binary_wasm_ts.js file in the build dir */
+var Binaryen = {};
+export default Binaryen;
